Complete Add menu ingredient

+++ PriceCalculator/app/src/main/java/com/example/pricecalculator/ArrayData/MenuIngredient.java
@@ -3,11 +3,11 @@ package com.example.pricecalculator.ArrayData;
 public class MenuIngredient {
     private String menu_ingredient_name;
     private int ingredient_id;
-    private double menu_ingredient_weight;
+    private int menu_ingredient_weight;
     private String menu_ingredient_unit;
     private double menu_ingredient_price;
 
-    public MenuIngredient(String menu_ingredient_name, int ingredient_id, double menu_ingredient_weight, String menu_ingredient_unit, double menu_ingredient_price) {
+    public MenuIngredient(String menu_ingredient_name, int ingredient_id, int menu_ingredient_weight, String menu_ingredient_unit, double menu_ingredient_price) {
         this.menu_ingredient_name = menu_ingredient_name;
         this.ingredient_id = ingredient_id;
         this.menu_ingredient_weight = menu_ingredient_weight;
@@ -23,7 +23,7 @@ public class MenuIngredient {
         this.menu_ingredient_name = menu_ingredient_name;
     }
 
-    public double getMenu_ingredient_weight() {
+    public int getMenu_ingredient_weight() {
         return menu_ingredient_weight;
     }
